+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added the `fs.addPublicExchangeKey()` function which adds the public exchange key of that domain/browser/device to your filesystem at `/public/.well-known/exchange/DID_KEY`. Along with the `fs.hasPublicExchangeKey()` to check if it's there.
 - Made the login low more resilient. Should work better with extensions triggering `postMessage`s now.
+- Updated keystore-idb to v0.15.0, which renamed `publicReadKey()` to `publicExchangeKey()` (among other functions). The read key-pair is now properly named the exchange key-pair.
+
 
 
 ### v0.27.0